Split ShelleyEra away from ShelleyBased
The ShelleyEra data type is now located in
Shelley.Spec.Ledger.ShelleyEra. The type family instances
have also been moved into this module.
