+++ shelley/chain-and-ledger/executable-spec/shelley-spec-ledger.cabal
@@ -52,6 +52,7 @@ library
     Shelley.Spec.Ledger.Serialization
     Shelley.Spec.Ledger.Slot
     Shelley.Spec.Ledger.SoftForks
+    Shelley.Spec.Ledger.ShelleyEra
     Shelley.Spec.Ledger.StabilityWindow
     Shelley.Spec.Ledger.STS.Bbody
     Shelley.Spec.Ledger.STS.Chain
