+++ shelley/chain-and-ledger/executable-spec/src/Cardano/Ledger/Shelley.hs
@@ -1,5 +1,4 @@
 {-# LANGUAGE ConstraintKinds #-}
-{-# LANGUAGE EmptyDataDecls #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE TypeFamilies #-}
@@ -8,22 +7,13 @@ module Cardano.Ledger.Shelley where
 
 import Cardano.Ledger.Compactible
 import Cardano.Ledger.Core
-import qualified Cardano.Ledger.Crypto as CryptoClass
 import Cardano.Ledger.Era
 import Cardano.Ledger.Val (Val)
-import Shelley.Spec.Ledger.Coin (Coin)
 import Shelley.Spec.Ledger.Hashing (EraIndependentTxBody, HashAnnotated (..))
 
 --------------------------------------------------------------------------------
 -- Shelley Era
 --------------------------------------------------------------------------------
-
-data ShelleyEra c
-
-instance CryptoClass.Crypto c => Era (ShelleyEra c) where
-  type Crypto (ShelleyEra c) = c
-
-type instance Value (ShelleyEra _) = Coin
 
 type TxBodyConstraints era =
   ( ChainData (TxBody era),
