+++ shelley/chain-and-ledger/executable-spec/src/Shelley/Spec/Ledger/API.hs
@@ -10,13 +10,14 @@ module Shelley.Spec.Ledger.API
 where
 
 import Cardano.Ledger.Era (Crypto)
-import Cardano.Ledger.Shelley (ShelleyBased, ShelleyEra)
+import Cardano.Ledger.Shelley (ShelleyBased)
 import Shelley.Spec.Ledger.API.ByronTranslation as X
 import Shelley.Spec.Ledger.API.Mempool as X
 import Shelley.Spec.Ledger.API.Protocol as X
 import Shelley.Spec.Ledger.API.Types as X
 import Shelley.Spec.Ledger.API.Validation as X
 import Shelley.Spec.Ledger.API.Wallet as X
+import Shelley.Spec.Ledger.ShelleyEra (ShelleyEra)
 
 class
   ( PraosCrypto (Crypto era),
